Fix plato menu

+++ Restaurant/src/main/java/com/uptc/frw/restaurant/controller/DishMenuController.java
@@ -38,7 +38,12 @@ public class DishMenuController {
         }
 
         DishMenuKey  dishMenuKey = new DishMenuKey(dish,menu);
-        return dishMenuService.findDishMenuById(dishMenuKey);
+        DishMenu existingDishMenu = dishMenuService.findDishMenuById(dishMenuKey);
+        if (existingDishMenu == null) {
+            throw new RuntimeException("La relación plato-menu no existe");
+        }
+
+        return existingDishMenu;
     }
 
     @PostMapping
@@ -61,6 +66,11 @@ public class DishMenuController {
         }
 
         DishMenuKey  dishMenuKey = new DishMenuKey(dish,menu);
+        DishMenu existingDishMenu = dishMenuService.findDishMenuById(dishMenuKey);
+        if (existingDishMenu == null) {
+            throw new RuntimeException("La relación plato-menu no existe");
+        }
+
         dishMenuService.deleteDishMenu(dishMenuKey);
     }
 }
